requirements.txt
# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Existing Dependencies
requests==2.31.0
pymupdf==1.24.0
langchain-text-splitters==0.3.0
numpy==1.26.0

# Utilities
python-dotenv==1.0.0
werkzeug==3.1.0
